Finishing makeDeal method

+++ src/com/company/repositories/interfaces/IDealRepository.java
@@ -12,7 +12,7 @@ public interface IDealRepository {
     boolean addCustomer(Customer customer); // add customers to the database
     boolean addProperty(Property property); // add property to the database
     boolean makeDeal(Deal deal); // make a deal
-    boolean addMoneyToBankAccount(int id, double money); // adding money to one's bank account
+    boolean addMoneyToBankAccount(int id, int money); // adding money to one's bank account
     boolean setDiscountForHouses(double discount); // setting discount for houses
     boolean setPublicServicesForFlats(double publicServices); // setting money for public services
     ArrayList<Customer> showCustomers(); // show all customers
